% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oh_compress}
\alias{oh_compress}
\title{Compresses a one-hot vector}
\usage{
oh_compress(onehot)
}
\arguments{
\item{A}{vector of one-hot values, maximum size 8 elements}
}
\value{
A compressed integer
}
\description{
Compresses a one-hot vector
}
