% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oh_decompress}
\alias{oh_decompress}
\title{Decompresses an integer to one-hot vector}
\usage{
oh_decompress(compressed, size)
}
\arguments{
\item{An}{integer (<8 bits wide) to decompress}
}
\value{
A one-hot vector
}
\description{
Decompresses an integer to one-hot vector
}
